fix: refactor document-pair validation function to use validateDocumentWithReferences
